Binary size reduction optimizations (athena-framework/athena#648)
## Context

As a result of #636, the `ParameterBag` type now included `ActionBase+`
in its union type. This ultimately made Crystal generate like 90+ more
`inspect` methods for many different Athena types. The exact cause of
this is a bit of a TBD, but best I can tell the gist of it is it's
because `.inspect` is being called on something that ends up cascading
to other types and at some point finds its way to the `value` of some
container types which then has to include all possible concrete types
stored within.

This PR breaks that cascade by overriding `#inspect` to not reference
the internal value that holds the variable type. Will see how this goes,
an always tweak the messages if needed. Also only need to do this for
structs since its default `#inspect` includes all the ivars, whereas
`class` does not do that.

This PR also makes `ATH::Action` a class to reduce the size of unions
that reference it as well.

## Changelog

- Define `#inspect` on internal container types to avoid referencing the
stored variable value
- Make `ATH::Action` a `class`